refactor: reorganize component files structure
- Move PlustButton to Views/PlusButton (rename PlustButton -> PlusButton)
- Move SearchableViewModifier to separate directory
- Move SkeletonView to Views directory for better organization
