--- Sclerosis/Screens/Components/SearchableViewModifier.swift
@@ -1,18 +1,0 @@
-//
-//  Searchable.swift
-//  Sclerosis
-//
-//  Created by Danil Kazakov on 25.09.2025.
-//
-
-import SwiftUI
-
-struct SearchableViewModifier: ViewModifier {
-    var body: some View {
-        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
-    }
-}
-
-#Preview {
-    SearchableViewModifier()
-}
